requirements.txt
# Core Dependencies
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
python-telegram-bot>=20.0
schedule>=1.2.0
python-dateutil>=2.8.0

# Logging and Utilities
colorlog>=6.7.0
colorama>=0.4.6

# Data Processing
pandas>=1.5.0
numpy>=1.24.0

# HTTP and Networking
httpx>=0.24.0
requests>=2.28.0

# Async Support (Windows compatibility)
uvloop>=0.17.0; sys_platform != "win32"

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Performance Monitoring
psutil>=5.9.0

# Optional: Environment Variables
python-dotenv>=1.0.0

# Optional: Database Support (if needed later)
# sqlalchemy>=2.0.0
# aiosqlite>=0.19.0

# Optional: Caching (if needed later)
# redis>=4.5.0
# aioredis>=2.0.0
